Added A/B looping
Changed to allow setting of the A mark and the B mark. If set playback
will loop between A and B. Since we are using the B key for Mark B, we
changed the f (forward) and b (backward) to be left and right arrow.

Also added the A and B marks under the player so you can see what you
have A and B set to.

Finally, updated question.mark.html to reflect changes.